Some succes with ioctl. Minimal version of proc driver (helloworld on read). Need to write open function for proc (ioctl cant work without open)

+++ modules/proc/helloworld.c
@@ -9,7 +9,7 @@ static int len, temp;
 
 static char *msg;
 
-int read_proc(struct file *filp, char *buf, size_t count, loff_t *offp) {
+ssize_t read_proc(struct file *filp, char *buf, size_t count, loff_t *offp) {
 
 	if(count > temp) {
 		count = temp;
